feat(react-draggable-dialog): split component logic into separate hooks
