+++ packages/react-draggable-dialog/src/components/DraggableDialogHandle/DraggableDialogHandle.tsx
@@ -1,20 +1,20 @@
 import * as React from 'react';
-import { useDraggable } from '@dnd-kit/core';
 import { mergeClasses } from '@fluentui/react-components';
 
 import { DraggableDialogHandleProps } from './DraggableDialogHandle.types';
 import { useStyles } from './DraggableDialogHandle.styles';
-import { useDraggableDialogState } from '../../contexts/DraggableDialogContext';
+import { useDraggableDialogHandle } from './useDraggableDialogHandle';
 
+/**
+ * DraggableDialogHandle is the handle element that can be used to drag the DraggableDialog.
+ */
 export const DraggableDialogHandle: React.FC<DraggableDialogHandleProps> = ({
   children,
   className,
 }) => {
   const styles = useStyles();
-  const { id } = useDraggableDialogState();
-  const { setActivatorNodeRef, attributes, listeners } = useDraggable({
-    id,
-  });
+  const { setActivatorNodeRef, attributes, listeners } =
+    useDraggableDialogHandle();
 
   return React.cloneElement(children, {
     ref: setActivatorNodeRef,
