+++ packages/react-draggable-dialog/src/components/DraggableDialogHandle/DraggableDialogHandle.types.ts
@@ -1,3 +1,4 @@
+import { SyntheticListenerMap } from '@dnd-kit/core/dist/hooks/utilities/useSyntheticListeners';
 import * as React from 'react';
 
 export type DraggableDialogHandleProps = React.HTMLAttributes<HTMLElement> & {
@@ -7,4 +8,21 @@ export type DraggableDialogHandleProps = React.HTMLAttributes<HTMLElement> & {
    * @required
    */
   children: React.ReactElement;
+};
+
+export type DraggableDialogHandleState = {
+  /**
+   * The ref to the activator node.
+   */
+  setActivatorNodeRef: React.RefCallback<HTMLElement>;
+
+  /**
+   * The attributes to apply to the activator node.
+   */
+  attributes: React.HTMLAttributes<HTMLElement>;
+
+  /**
+   * The event listeners to apply to the activator node.
+   */
+  listeners?: SyntheticListenerMap;
 };
