+++ packages/react-draggable-dialog/src/components/DraggableDialogSurface/DraggableDialogSurface.types.ts
@@ -1,3 +1,15 @@
 import { DialogSurfaceProps } from '@fluentui/react-components';
 
 export type DraggableDialogSurfaceProps = DialogSurfaceProps;
+
+export type DraggableDialogSurfaceState = {
+  /**
+   * Ref to the draggable dialog surface.
+   */
+  ref: React.RefObject<HTMLDivElement>;
+
+  /**
+   * Style to apply to the draggable dialog surface.
+   */
+  style: React.CSSProperties | undefined;
+};
